backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.9.0
